Disable logging by default
Since `notify-send.sh` doesn't run as a service but on invocation only, the only purpose logging serves is debugging the script's usage (not necessarily the script itself, pass the `DEBUG=true` env variable for that). Permanently enabling logging causes `notify-send.sh` to constantly create logfiles that serve no purpose unless the user is actively trying to debug the script's usage. Thus it's better to enable logging only when logging was requested by the user. To enable logging, pass the `LOGGING=true` env variable.

Another reason to disable logging by default is that it breaks assigning the script's output (e.g. the notification ID with `--print-id`) to a variable on some platforms.